created new file for state level maps
